+++ portfolio/src/components/main/Projects.tsx
@@ -3,10 +3,7 @@ import ProjectCard from '../sub/ProjectCard';
 
 const Projects = () => {
   return (
-    <div
-      className='flex flex-col items-center justify-center py-20'
-      id='projects'
-    >
+    <div className='flex flex-col items-center justify-center py-20'>
       <h1 className='text-[40px] font-semibold text-transparent bg-clip-text bg-gradient-to-r from-purple-500 to-cyan-500 py-20'>
         My Projects
       </h1>
