+++ portfolio/src/components/resume/ResumeCard.js
@@ -10,7 +10,7 @@ const ResumeCard = ({ title, subTitle, result, des }) => {
           <span className='w-3 h-3 rounded-full bg-bodyColor inline-flex group-hover:bg-designColor duration-300'></span>
         </span>
       </div>
-      <div className='w-full p-[4em] bg-black bg-opacity-20 hover:bg-opacity-30 duration-300  rounded-lg  lgl:px-10 flex flex-col justify-center gap-6 lgl:gap-10 shadow-shadowOne'>
+      <div className='w-[17rem] p-[0.5rem] md:w-full  md:p-[4em] bg-black bg-opacity-20 hover:bg-opacity-30 duration-300  rounded-lg  lgl:px-10 flex flex-col justify-center gap-6 lgl:gap-10 shadow-shadowOne'>
         <div className='flex flex-col lgl:flex-row justify-between gap-4 lgl:gap-0 lgl:items-center'>
           <div>
             <h3 className='text-xl md:text-2xl font-semibold group-hover:text-white duration-300 text-lightText'>
